Status Bars and Consumption

+++ Assets/Scripts/StatusBars.cs
@@ -10,37 +10,37 @@ public class StatusBars : MonoBehaviour
     public Slider thirstSlider;
 
     //Health Status Bar//
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         healthSlider.maxValue = health;
         healthSlider.value = health;
     }
     
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         healthSlider.value = health;
     }
 
     //Hunger Status Bar//
-    public void SetMaxHunger(int hunger)
+    public void SetMaxHunger(float hunger)
     {
         hungerSlider.maxValue = hunger;
         hungerSlider.value = hunger;
     }
 
-    public void SetHunger(int hunger)
+    public void SetHunger(float hunger)
     {
         hungerSlider.value = hunger;
     }
 
     //Thirst Status Bar//
-    public void SetMaxThirst(int health)
+    public void SetMaxThirst(float health)
     {
         thirstSlider.maxValue = health;
         thirstSlider.value = health;
     }
 
-    public void SetThirst(int thirst)
+    public void SetThirst(float thirst)
     {
         thirstSlider.value = thirst;
     }
